+++ database/migrations/2018_10_19_002948_create_servicios_empresa_table.php
@@ -18,7 +18,7 @@ class CreateServiciosEmpresaTable extends Migration
             $table->string('nombreEmpresa');
             $table->string('tipoEmpresa');
             $table->integer('numAccionistas');
-            $table->integer('capital');
+            $table->bigInteger('capital');
             $table->string('departamento');
             $table->string('ciudad');
             $table->integer('FK_UsuarioId')->unsigned();
